Split the music service browser into a package and drop the unused favorites API
